requirements.txt
numpy >= 1.9.0
scipy
singledispatch >= 3.4.0.3
scikit-learn >= 0.18
attrs > 16.0.0
jinja2
pip >= 8.1
setuptools >= 20.7
